+++ CHANGELOG.md
@@ -58,6 +58,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * feat: when dumping pane viewport, allow including the styles (ansi escape codes) for color rendering in editors that support it (https://github.com/zellij-org/zellij/pull/4617)
 * fix: tab-bar wrapping for extremely small terminal widths (https://github.com/zellij-org/zellij/pull/4728)
 * refactor: replace `async_std` with `tokio` (https://github.com/zellij-org/zellij/pull/4672)
+* fix: do not attempt a reconnect from the web-client when explicitly kicked from the session (https://github.com/zellij-org/zellij/pull/4730)
 
 
 ## [0.43.1] - 2025-08-08